[rocRoller] Add experimental option to omit permlane instructions. (#737)
Turning this branch from @pjots into a kernel option. Validation will
fail if this option is enabled but this will allow the code path to not
fall behind.

## To this end (@pjots' work):

- Added new `Segment` DataFlow edge which selects a particular sub-dword
value from a packed type.
- `ControlFlowRWTracer` deals with Segment and Index edges generally
instead of individually for each node type.

## Other changes:

- It is now an error to add a register with a negative tag value to
`RegisterTagManager`.
    - Got rid of a few cases in lib and tests where this was happening.

### `m0` calculation is now done entirely using scalar ALU
- Scalar offset value for 
- Add new `ToScalar` expression and `makeScalar` expression
transformation to shift `readfirstlane` to values instead of at the end
of the expression
    -  [ ] TODO: Move this to an `ArithmeticGenerator`
    -  [ ] TODO: Add a test for makeScalar transformation.
    -  [ ] TODO: Move makeScalar to its own CPP file.
- This also gets rid of some temporary values and `mov` instructions.
### GEMM Client changes
- Deduplicated some of the GEMM client struct fields.
- Refactor GEMM client serialization

---------

+++ shared/rocroller/lib/include/rocRoller/InstructionValues/Register.hpp
@@ -337,7 +337,7 @@ namespace rocRoller
              *  - bitWidth is zero or greater than number of bits in a register; and/or
              *  - the bitfield would cross register boundaries.
              */
-            ValuePtr bitfield(uint8_t bitOffset, uint8_t bitWidth) const;
+            ValuePtr bitfield(int bitOffset, int bitWidth) const;
 
             /**
              * Return contiguous segments of a packed datatype denotated by indices.
@@ -359,13 +359,13 @@ namespace rocRoller
              * Returns the bit offset of this bitfield Value.
              * Throws FatalError if this is not a bitfield Value.
              */
-            uint8_t getBitOffset() const;
+            int getBitOffset() const;
 
             /**
              * Returns the bit width (length in bits) of this bitfield.
              * Throws FatalError if this is not a bitfield Value.
              */
-            uint8_t getBitWidth() const;
+            int getBitWidth() const;
 
         private:
             /**
@@ -406,8 +406,8 @@ namespace rocRoller
              * Offset and bit-width of the bitfielf in the 32-bit register represented by this value.
              * Multi-register values must have empty bitOffset & bitWidth, otherwise they are invalid.
              */
-            std::optional<uint8_t> m_bitOffset;
-            std::optional<uint8_t> m_bitWidth;
+            std::optional<int> m_bitOffset;
+            std::optional<int> m_bitWidth;
 
             /**
              * Pulls values from the allocation
